+++ src/main/resources/import.sql
@@ -10,11 +10,11 @@ insert into vehicles(PLATENO,CAR_MODEL,YEAR,COLOR,USER_ID) values ('IMA-8354','A
 insert into vehicles(PLATENO,CAR_MODEL,YEAR,COLOR,USER_ID) values ('IMA-8353','AUDI','2007','GREY','3');
 insert into vehicles(PLATENO,CAR_MODEL,YEAR,COLOR,USER_ID) values ('IMA-8352','AUDI','2007','GREY','4');
 
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/11','11:15','Completed','Small','500','1');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/11','09:30','Completed','Big','1000','2');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/11','11:15','Completed','Small','500','1');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/11','09:30','Completed','Big','1000','2');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/27','11:15','In_Progress','Small','500','1');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/27','12:30','In_Progress','Big','1000','2');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/27','13:15','On_Hold','Small','500','3');
-insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('2017/10/27','15:30','On_Hold','Big','1000','4');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('11-10-2017','11:15','Completed','Small','500','1');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('11/10/2017','09:30','Completed','Big','1000','2');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('11/10/2017','11:15','Completed','Small','500','1');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('11/10/2017','09:30','Completed','Big','1000','2');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('27/10/2017','11:15','In_Progress','Small','500','1');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('27/10/2017','12:30','In_Progress','Big','1000','2');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('27/10/2017','13:15','On_Hold','Small','500','3');
+insert into repairs(REPAIR_DATE,REPAIR_TIME,REPAIR_STATUS,REPAIR_TYPE,REPAIR_COST,VEHICLE_ID) values ('27/10/2017','15:30','On_Hold','Big','1000','4');
